add homepage links
