+++ Sass/base/_typography.scss
@@ -1,4 +1,3 @@
-//global font we decided
 @import url('https://fonts.googleapis.com/css2?family=Roboto:wght@300&display=swap');
 
 $roboto: 'Roboto', sans-serif;
