+++ src/Home/JournalCard.js
@@ -1,18 +1,25 @@
-// import React, { useEffect, useState } from 'react'
-// import { useHistory } from 'react-router-dom';
+import React, { useEffect, useState } from 'react'
+import { useHistory } from 'react-router-dom';
 
+function JournalCard (props) {
 
-// function JournalCard (props) {
-//   return (
-//     <div className="journalSession" onClick={() => {handleClick()}}>
-//       <div className="homeSessionText">
-//           {`TODAY'S JOURNAL`}
-//         <div className="journalDisplay">
-//           {journal? 'YOU COMPLETED YOUR DAILY JOURNAL' : 'YOU HAVE NOT JOURNALED TODAY'}
-//         </div>
-//       </div>
-//     </div>
-//   )
-// }
+  const [journal, setJournal] = useState(false)
+  const history = useHistory()
 
-// export default JournalCard;
+  const handleClick = () => {
+    history.push('/journal')
+  }
+
+  return (
+    <div className="journalSession" onClick={() => {handleClick()}}>
+      <div className="homeSessionText">
+          {`TODAY'S JOURNAL`}
+        <div className="journalDisplay">
+          {journal? 'YOU COMPLETED YOUR DAILY JOURNAL' : 'YOU HAVE NOT JOURNALED TODAY'}
+        </div>
+      </div>
+    </div>
+  )
+}
+
+export default JournalCard;
