+++ src/Home/MealTile.js
@@ -1,7 +1,14 @@
 import React, { useEffect, useState } from 'react'
+import { useHistory } from 'react-router-dom';
 
 function MealTile (props) {
 
+  const history = useHistory();
+
+  const mealTileClick = () => {
+   console.log('mealTileclick');
+   history.push('/meals')
+ }
 
   return (
     <div className="mealTile" onClick={() => mealTileClick()}>
