+++ src/Home/MealsCard.js
@@ -5,6 +5,12 @@ import MealTile from './MealTile'
 
 function MealsCard (props) {
 
+  const history = useHistory();
+
+   const mealTileClick = () => {
+    console.log('mealTileclick');
+    history.push('/meals')
+  }
 
   return (
     <div className="mealSession" >
